+++ code/test/sixsq/nuvla/server/resources/nuvlabox_1_lifecycle_test.clj
@@ -681,7 +681,7 @@
                              (ltu/is-key-value :state "ACTIVATED")
                              (ltu/get-op-url :commission))]
 
-          ;; commissioning of the nuvlabox (no swarm credentials)
+          ;; commissioning of the nuvlabox
           (-> session-owner
               (request commission
                        :request-method :post
@@ -817,6 +817,235 @@
                 (if (= "kubernetes" subtype)
                   (is (= 1 (count creds))))
                 )))
+
+          )))))
+
+
+(deftest create-activate-commission-removed-lifecycle
+  (binding [config-nuvla/*stripe-api-key* nil]
+    (let [session       (-> (ltu/ring-app)
+                            session
+                            (content-type "application/json"))
+
+          session-owner (header session authn-info-header "user/alpha group/nuvla-user group/nuvla-anon")
+          session-anon  (header session authn-info-header "unknown group/nuvla-anon")]
+
+      (let [nuvlabox-id      (-> session-owner
+                                 (request base-uri
+                                          :request-method :post
+                                          :body (json/write-str valid-nuvlabox))
+                                 (ltu/body->edn)
+                                 (ltu/is-status 201)
+                                 (ltu/location))
+
+            session-nuvlabox (header session authn-info-header
+                                     (str nuvlabox-id
+                                          " group/nuvla-nuvlabox group/nuvla-anon"))
+
+            nuvlabox-url     (str p/service-context nuvlabox-id)
+
+            activate-url     (-> session-owner
+                                 (request nuvlabox-url)
+                                 (ltu/body->edn)
+                                 (ltu/is-status 200)
+                                 (ltu/is-operation-present :edit)
+                                 (ltu/is-operation-present :delete)
+                                 (ltu/is-operation-present :activate)
+                                 (ltu/is-operation-absent :commission)
+                                 (ltu/is-operation-absent :decommission)
+                                 (ltu/is-key-value :state "NEW")
+                                 (ltu/get-op-url :activate))]
+
+        ;; activate nuvlabox
+        (-> session-anon
+            (request activate-url
+                     :request-method :post)
+            (ltu/body->edn)
+            (ltu/is-status 200)
+            (ltu/is-key-value (comp not str/blank?) :secret-key true)
+            (ltu/body)
+            :api-key
+            (ltu/href->url))
+
+        (let [{isg-id :id} (-> session-owner
+                               (content-type "application/x-www-form-urlencoded")
+                               (request isg-collection-uri
+                                        :request-method :put
+                                        :body (rc/form-encode {:filter (format "parent='%s'"
+                                                                               nuvlabox-id)}))
+                               (ltu/body->edn)
+                               (ltu/is-status 200)
+                               (ltu/is-count 1)
+                               (ltu/entries)
+                               first)
+
+              commission (-> session-owner
+                             (request nuvlabox-url)
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/is-operation-present :edit)
+                             (ltu/is-operation-absent :delete)
+                             (ltu/is-operation-absent :activate)
+                             (ltu/is-operation-present :commission)
+                             (ltu/is-operation-present :decommission)
+                             (ltu/is-key-value :state "ACTIVATED")
+                             (ltu/get-op-url :commission))]
+
+          ;; commissioning of the nuvlabox
+          (-> session-nuvlabox
+              (request commission
+                       :request-method :post
+                       :body (json/write-str {:swarm-token-worker     "abc"
+                                              :swarm-token-manager    "def"
+                                              :swarm-client-key       "key"
+                                              :swarm-client-cert      "cert"
+                                              :swarm-client-ca        "ca"
+                                              :swarm-endpoint         "https://swarm.example.com"
+                                              :minio-access-key       "access"
+                                              :minio-secret-key       "secret"
+                                              :minio-endpoint         "https://minio.example.com"
+                                              :kubernetes-client-key  "key"
+                                              :kubernetes-client-cert "cert"
+                                              :kubernetes-client-ca   "ca"
+                                              :kubernetes-endpoint    "https://k8s.example.com"}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
+
+          ;; verify state of the resource
+          (-> session-owner
+              (request nuvlabox-url)
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/is-operation-present :edit)
+              (ltu/is-operation-absent :delete)
+              (ltu/is-operation-absent :activate)
+              (ltu/is-operation-present :commission)
+              (ltu/is-operation-present :decommission)
+              (ltu/is-key-value :state "COMMISSIONED"))
+
+          ;; check that services exist
+          (let [services (-> session-owner
+                             (content-type "application/x-www-form-urlencoded")
+                             (request infra-service-collection-uri
+                                      :request-method :put
+                                      :body (rc/form-encode {:filter (format
+                                                                       "parent='%s'" isg-id)}))
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/is-count 3)
+                             (ltu/entries))]
+
+            (is (= #{"swarm" "s3" "kubernetes"} (set (map :subtype services))))
+
+            (doseq [{:keys [acl]} services]
+              (is (= [nuvlabox-owner] (:view-acl acl))))
+
+            (doseq [{:keys [subtype] :as service} services]
+              (let [creds (-> session-owner
+                              (content-type "application/x-www-form-urlencoded")
+                              (request credential-collection-uri
+                                       :request-method :put
+                                       :body (rc/form-encode {:filter (format "parent='%s'"
+                                                                              (:id service))}))
+                              (ltu/body->edn)
+                              (ltu/is-status 200)
+                              (ltu/entries))]
+
+                (if (= "swarm" subtype)
+                  (is (= 3 (count creds))))                 ;; only swarm token credentials
+
+                (if (= "s3" subtype)
+                  (is (= 1 (count creds))))                 ;; only key/secret pair
+
+                (if (= "kubernetes" subtype)
+                  (is (= 1 (count creds))))
+                )))
+
+          ;; commissioning with removed
+          (-> session-nuvlabox
+              (request commission
+                       :request-method :post
+                       :body (json/write-str {:swarm-token-worker     "abc"
+                                              :swarm-token-manager    "def"
+                                              :swarm-client-key       "key"
+                                              :swarm-client-cert      "cert"
+                                              :swarm-client-ca        "ca"
+                                              :swarm-endpoint         "https://swarm.example.com"
+                                              :minio-access-key       "access"
+                                              :minio-secret-key       "secret"
+                                              :minio-endpoint         "https://minio.example.com"
+                                              :kubernetes-client-key  "key"
+                                              :kubernetes-client-cert "cert"
+                                              :kubernetes-client-ca   "ca"
+                                              :kubernetes-endpoint    "https://k8s.example.com"
+                                              :removed                ["swarm-endpoint"
+                                                                       "swarm-token-manager"
+                                                                       "swarm-token-worker"
+                                                                       "swarm-client-key"]}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
+
+          ;; check that services exist
+          (let [services (-> session-owner
+                             (content-type "application/x-www-form-urlencoded")
+                             (request infra-service-collection-uri
+                                      :request-method :put
+                                      :body (rc/form-encode {:filter (format
+                                                                       "parent='%s'" isg-id)}))
+                             (ltu/body->edn)
+                             (ltu/is-status 200)
+                             (ltu/is-count 2)
+                             (ltu/entries))]
+
+            (is (= #{"s3" "kubernetes"} (set (map :subtype services))))
+
+            (doseq [{:keys [acl]} services]
+              (is (= [nuvlabox-owner] (:view-acl acl))))
+
+            (doseq [{:keys [subtype] :as service} ["swarm" "s3" "kubernetes"]]
+              (let [creds (-> session-owner
+                              (content-type "application/x-www-form-urlencoded")
+                              (request credential-collection-uri
+                                       :request-method :put
+                                       :body (rc/form-encode {:filter (format "parent='%s'"
+                                                                              (:id service))}))
+                              (ltu/body->edn)
+                              (ltu/is-status 200)
+                              (ltu/entries))]
+
+                (if (= "swarm" subtype)
+                  (is (= 0 (count creds))))                 ;; swarm creds all gone
+
+                (if (= "s3" subtype)
+                  (is (= 1 (count creds))))                 ;; only key/secret pair
+
+                (if (= "kubernetes" subtype)
+                  (is (= 1 (count creds))))
+                )))
+
+          ;; additionnal call to commission will not fail
+          (-> session-nuvlabox
+              (request commission
+                       :request-method :post
+                       :body (json/write-str {:swarm-token-worker     "abc"
+                                              :swarm-token-manager    "def"
+                                              :swarm-client-key       "key"
+                                              :swarm-client-cert      "cert"
+                                              :swarm-client-ca        "ca"
+                                              :swarm-endpoint         "https://swarm.example.com"
+                                              :minio-access-key       "access"
+                                              :minio-secret-key       "secret"
+                                              :minio-endpoint         "https://minio.example.com"
+                                              :kubernetes-client-key  "key"
+                                              :kubernetes-client-cert "cert"
+                                              :kubernetes-client-ca   "ca"
+                                              :kubernetes-endpoint    "https://k8s.example.com"
+                                              :removed                ["swarm-endpoint"
+                                                                       "swarm-token-manager"
+                                                                       "swarm-token-worker"
+                                                                       "swarm-client-key"]}))
+              (ltu/body->edn)
+              (ltu/is-status 200))
 
           )))))
 
@@ -1216,7 +1445,7 @@
               (ltu/body->edn)
               (ltu/is-status 200)
               (ltu/is-key-value :execution-mode "pull")
-              (ltu/is-key-value :acl {:edit-data ["user/alpha"],
+              (ltu/is-key-value :acl {:edit-data [nuvlabox-id "user/alpha"],
                                       :owners    ["group/nuvla-admin"],
                                       :view-acl  ["user/alpha"],
                                       :delete    ["user/alpha"],
@@ -1224,7 +1453,7 @@
                                       :edit-acl  ["user/alpha"],
                                       :view-data [nuvlabox-id "user/alpha"],
                                       :manage    [nuvlabox-id "user/alpha"],
-                                      :edit-meta ["user/alpha"]}))))
+                                      :edit-meta [nuvlabox-id "user/alpha"]}))))
 
       )))